chore(bindings): regenerate after Swift codegen fixes (keyword escape, phantom Vec)
Picks up alef's Swift backend keyword-escape fixes (subscript_, extension_)
and the phantom Vec<FooBox> entries in the trait bridge extern blocks
that force swift-bridge-build to emit the Vec C symbols. Also splits
RustBridgeC umbrella header to include SwiftBridgeCore.h + kreuzberg-swift.h
properly.
