+++ packages/swift/Sources/RustBridgeC/kreuzberg-swift.h
@@ -2740,8 +2740,8 @@ bool __swift_bridge__$TreeSitterProcessConfig$symbols(void* self);
 bool __swift_bridge__$TreeSitterProcessConfig$diagnostics(void* self);
 struct __private__OptionUsize __swift_bridge__$TreeSitterProcessConfig$chunk_max_size(void* self);
 void* __swift_bridge__$TreeSitterProcessConfig$content_mode(void* self);
-void* __swift_bridge__$SupportedFormat$new(void* extension, void* mime_type);
-void* __swift_bridge__$SupportedFormat$extension(void* self);
+void* __swift_bridge__$SupportedFormat$new(void* extension_, void* mime_type);
+void* __swift_bridge__$SupportedFormat$extension_(void* self);
 void* __swift_bridge__$SupportedFormat$mime_type(void* self);
 void* __swift_bridge__$ServerConfig$new(void* host, uint16_t port, void* cors_origins, uintptr_t max_request_body_bytes, uintptr_t max_multipart_field_bytes);
 void* __swift_bridge__$ServerConfig$host(void* self);
@@ -3680,6 +3680,7 @@ struct __private__ResultPtrAndPtr __swift_bridge__$render_pdf_page_to_png(void* 
 struct __private__ResultPtrAndPtr __swift_bridge__$extract_text_from_pdf(void* pdf_bytes);
 struct __private__ResultPtrAndPtr __swift_bridge__$serialize_to_toon(void* result);
 struct __private__ResultPtrAndPtr __swift_bridge__$serialize_to_json(void* result);
+void* __swift_bridge__$_alef_phantom_vec_ocr_backend(void);
 struct __private__ResultPtrAndPtr __swift_bridge__$ocr_backend_call_process_image(void* this, void* image_bytes, void* config);
 struct __private__ResultPtrAndPtr __swift_bridge__$ocr_backend_call_process_image_file(void* this, void* path, void* config);
 bool __swift_bridge__$ocr_backend_call_supports_language(void* this, void* lang);
@@ -3688,13 +3689,16 @@ void* __swift_bridge__$ocr_backend_call_supported_languages(void* this);
 bool __swift_bridge__$ocr_backend_call_supports_table_detection(void* this);
 bool __swift_bridge__$ocr_backend_call_supports_document_processing(void* this);
 struct __private__ResultPtrAndPtr __swift_bridge__$ocr_backend_call_process_document(void* this, void* path, void* config);
+void* __swift_bridge__$_alef_phantom_vec_post_processor(void);
 void* __swift_bridge__$post_processor_call_process(void* this, void* result, void* config);
 void* __swift_bridge__$post_processor_call_processing_stage(void* this);
 bool __swift_bridge__$post_processor_call_should_process(void* this, void* result, void* config);
 uint64_t __swift_bridge__$post_processor_call_estimated_duration_ms(void* this, void* result);
+void* __swift_bridge__$_alef_phantom_vec_validator(void);
 void* __swift_bridge__$validator_call_validate(void* this, void* result, void* config);
 bool __swift_bridge__$validator_call_should_validate(void* this, void* result, void* config);
 int32_t __swift_bridge__$validator_call_priority(void* this);
+void* __swift_bridge__$_alef_phantom_vec_embedding_backend(void);
 uintptr_t __swift_bridge__$embedding_backend_call_dimensions(void* this);
 struct __private__ResultPtrAndPtr __swift_bridge__$embedding_backend_call_embed(void* this, void* texts);
