+++ packages/swift/rust/src/lib.rs
@@ -522,8 +522,8 @@ mod ffi {
     extern "Rust" {
         type SupportedFormat;
         #[swift_bridge(init)]
-        fn new(extension: String, mime_type: String) -> SupportedFormat;
-        fn extension(&self) -> String;
+        fn new(extension_: String, mime_type: String) -> SupportedFormat;
+        fn extension_(&self) -> String;
         fn mime_type(&self) -> String;
     }
 
@@ -2981,6 +2981,7 @@ mod ffi {
 
     extern "Rust" {
         type OcrBackendBox;
+        fn _alef_phantom_vec_ocr_backend() -> Vec<OcrBackendBox>;
         fn ocr_backend_call_process_image(
             this: &OcrBackendBox,
             image_bytes: Vec<u8>,
@@ -3005,6 +3006,7 @@ mod ffi {
 
     extern "Rust" {
         type PostProcessorBox;
+        fn _alef_phantom_vec_post_processor() -> Vec<PostProcessorBox>;
         fn post_processor_call_process(
             this: &PostProcessorBox,
             result: ExtractionResult,
@@ -3021,6 +3023,7 @@ mod ffi {
 
     extern "Rust" {
         type ValidatorBox;
+        fn _alef_phantom_vec_validator() -> Vec<ValidatorBox>;
         fn validator_call_validate(
             this: &ValidatorBox,
             result: ExtractionResult,
@@ -3036,6 +3039,7 @@ mod ffi {
 
     extern "Rust" {
         type EmbeddingBackendBox;
+        fn _alef_phantom_vec_embedding_backend() -> Vec<EmbeddingBackendBox>;
         fn embedding_backend_call_dimensions(this: &EmbeddingBackendBox) -> usize;
         fn embedding_backend_call_embed(this: &EmbeddingBackendBox, texts: Vec<String>) -> Result<String, String>;
     }
@@ -4753,10 +4757,10 @@ impl TreeSitterProcessConfig {
 pub struct SupportedFormat(pub kreuzberg::SupportedFormat);
 
 impl SupportedFormat {
-    pub fn new(extension: String, mime_type: String) -> SupportedFormat {
+    pub fn new(extension_: String, mime_type: String) -> SupportedFormat {
         ::std::unimplemented!("constructor not available: struct requires Default which is not implemented")
     }
-    pub fn extension(&self) -> String {
+    pub fn extension_(&self) -> String {
         serde_json::to_string(&self.0.extension).unwrap_or_default()
     }
     pub fn mime_type(&self) -> String {
@@ -11904,6 +11908,11 @@ pub fn serialize_to_json(result: ExtractionResult) -> Result<String, String> {
 
 pub struct OcrBackendBox(pub Box<dyn kreuzberg::plugins::OcrBackend + Send + Sync>);
 
+#[doc(hidden)]
+pub fn _alef_phantom_vec_ocr_backend() -> Vec<OcrBackendBox> {
+    Vec::new()
+}
+
 pub fn ocr_backend_call_process_image(
     this: &OcrBackendBox,
     image_bytes: Vec<u8>,
@@ -11980,6 +11989,11 @@ pub fn ocr_backend_call_process_document(
 
 pub struct PostProcessorBox(pub Box<dyn kreuzberg::plugins::PostProcessor + Send + Sync>);
 
+#[doc(hidden)]
+pub fn _alef_phantom_vec_post_processor() -> Vec<PostProcessorBox> {
+    Vec::new()
+}
+
 pub fn post_processor_call_process(
     this: &PostProcessorBox,
     mut result: ExtractionResult,
@@ -12015,6 +12029,11 @@ pub fn post_processor_call_estimated_duration_ms(this: &PostProcessorBox, result
 
 pub struct ValidatorBox(pub Box<dyn kreuzberg::plugins::Validator + Send + Sync>);
 
+#[doc(hidden)]
+pub fn _alef_phantom_vec_validator() -> Vec<ValidatorBox> {
+    Vec::new()
+}
+
 pub fn validator_call_validate(
     this: &ValidatorBox,
     result: ExtractionResult,
@@ -12036,6 +12055,11 @@ pub fn validator_call_priority(this: &ValidatorBox) -> i32 {
 }
 
 pub struct EmbeddingBackendBox(pub Box<dyn kreuzberg::plugins::EmbeddingBackend + Send + Sync>);
+
+#[doc(hidden)]
+pub fn _alef_phantom_vec_embedding_backend() -> Vec<EmbeddingBackendBox> {
+    Vec::new()
+}
 
 pub fn embedding_backend_call_dimensions(this: &EmbeddingBackendBox) -> usize {
     this.0.dimensions()
